requirements.txt
streamlit
strands-agents
strands-agents[ollama]
strands-agents-tools
mcp
plotly
pandas
numpy
requests
